+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## cycler v0.2.0
 - Removed commented out code
+- Added more default build functions
 
 ## cycler v0.1.3
 - Added github actions to verify with clippy and miri